dms/requirements.txt
# pip install -r requirements.txt

psycopg
transformers
torch
pillow
pdf2image
pytesseract

uvicorn
fastapi
sqlalchemy
alembic

passlib[argon2]
python-jose[cryptography]
pydantic[email]

python-multipart

# Optional integrations (disabled until code paths are uncommented):
# boto3
# botocore
# minio
# opencv-python-headless
# numpy
# requests
# scikit-learn
